+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 ## [3.1.0] BREAKING CHANGE
 
 * API improved
-
 * `FormKeyboardActions` was renamed to `KeyboardActions`.
 * `KeyboardCustomInput` was added to help you to create custom keyboards in an easy way.
 * added `enabled` property inside `KeyboardActions` in case you don't want to use `KeyboardActions` widget (tablets for example).
 * added `displayActionBar` property inside `KeyboardAction` in case you want to display/hide the keyboard bar (E.g: if you use footerBuilder and add your own done button inside that)
+* added `isDialog` property inside `KeyboardActions`.
 * Material color is transparent to avoid issues with the parent container.
-* Note: custom keyboards still doesn't work well on Dialog
+
 
 
 ## [3.0.0] BREAKING CHANGE